% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellESet-class}
\alias{CellESet-class}
\alias{CellESet}
\title{class definition for class CellESet}
\description{
class definition for class CellESet
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{Raw matrix for expression profile.}

\item{\code{meta.data}}{Data frame for storing the meta information of cells.}

\item{\code{use.cells}}{Cells used for prediction. Default is all cells.}

\item{\code{version}}{package version}
}}

