% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{ImportData}
\alias{ImportData}
\title{Import expression profiles from other object as object of class CellESet}
\usage{
ImportData(object)
}
\arguments{
\item{object}{An object of class \code{seurat}, \code{Seurat}, \code{matrix}(can be sparse) or \code{data.frame}
in which rows represent for features and columns for cells.}
}
\value{
An object of class CellESet(Cell Expression Set) with raw data stored in slot \code{raw.data}.
}
\description{
Import expression profiles from other object as object of class CellESet
}
\examples{
\dontrun{
library(Seurat)
myces <- ImportData(pbmc_small)
myces
}
}
