% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHist}
\alias{plotHist}
\title{plot histogram of predicted cell types}
\usage{
plotHist(object)
}
\arguments{
\item{object}{CellESet object.}
}
\value{
A ggplot object
}
\description{
plot histogram of predicted cell types
}
\examples{
\dontrun{
table(myces@meta.data$pred_types)
plotHist(myces)
}
}
