% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,CellESet-method}
\alias{show,CellESet-method}
\title{show method}
\usage{
\S4method{show}{CellESet}(object)
}
\arguments{
\item{object}{A CellESet object}
}
\description{
show method
}
