% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trans_dge}
\alias{trans_dge}
\title{prepare the expression profile and complete it if there's genes not found}
\usage{
trans_dge(mat, homogenes, species = "human")
}
\arguments{
\item{mat}{features x barcodes matrix}

\item{homogenes}{homogenes}

\item{species}{Either 'human' or 'mouse' for now.}
}
\value{
a (0,1)-matrix
}
\description{
prepare the expression profile and complete it if there's genes not found
}
